Improve Java SDK dependency
Current behavior:
The current Depends: field causes apt to install a JDK that might not be the default-jdk when the user has no JDK installed. On Ubuntu Xenial and Yaketty this will cause OpenJDK 9 to be installed, which is not optimal given that OpenJDK 9 has not been released yet.

Expected result:
Installing android-studio should cause the distro's recommended JDK (ie. default-jdk) to be installed when the user has no JDK installed. If the user already has a JDK installed, then android-studio should use it. This can be achieved by adding "default-jdk" to the "Depends:" field while still keeping the java-sdk dependency for when the user has another JDK installed.